+++ package.json
@@ -35,6 +35,7 @@
     "craco-less": "^1.16.0",
     "date-fns": "^2.8.1",
     "dotenv": "^8.1.0",
+    "eslint-plugin-flowtype": "^4.6.0",
     "exenv": "^1.2.2",
     "express": "^4.17.1",
     "express-http-proxy": "^1.6.0",
@@ -87,7 +88,7 @@
     "react-intl": "^3.3.1",
     "react-modal": "^3.11.1",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.0",
+    "react-scripts": "^3.4.0",
     "styled-components": "^4.4.0",
     "uuid": "^3.4.0"
   },
