+++ src/components/spørsmål/CheckboxSpørsmål.tsx
@@ -1,5 +1,5 @@
 import React, { SyntheticEvent } from 'react';
-import { ISpørsmål, ISvar } from '../../models/spørsmal';
+import { ISpørsmål, ISvar } from '../../models/spørsmålogsvar';
 import { useIntl } from 'react-intl';
 import { Element } from 'nav-frontend-typografi';
 import { CheckboksPanel } from 'nav-frontend-skjema';
