+++ src/helpers/standardSvar.ts
@@ -1,0 +1,6 @@
+import { ESvar } from '../models/spørsmålogsvar';
+
+export const JaNeiSvar = [
+  { svar_tekstid: ESvar.JA },
+  { svar_tekstid: ESvar.NEI },
+];
