+++ src/models/dokumentasjon.ts
@@ -1,0 +1,9 @@
+import { IVedlegg } from './vedlegg';
+
+export interface IDokumentasjon {
+  id: string;
+  tittel: string;
+  beskrivelse: string;
+  harSendtInn: boolean;
+  opplastedeVedlegg?: IVedlegg[];
+}
