+++ src/models/hjelpetekst.ts
@@ -1,0 +1,5 @@
+export interface IHjelpetekst {
+  åpneTekstid: string;
+  lukkeTekstid: string;
+  innholdTekstid: string;
+}
