+++ src/models/periode.ts
@@ -1,0 +1,6 @@
+import { IDatoFelt } from './søknadsfelter';
+
+export interface IPeriode {
+  fra: IDatoFelt;
+  til: IDatoFelt;
+}
