+++ src/models/steg/aktivitet/arbeidsgiver.ts
@@ -1,7 +1,7 @@
 import { IBooleanFelt, IDatoFelt, ITekstFelt } from '../../søknadsfelter';
 
 export interface IArbeidsgiver {
-  react_key: string;
+  id: string;
   navn?: ITekstFelt;
   arbeidsmengde?: ITekstFelt;
   fastStilling?: ITekstFelt;
