+++ src/søknad/steg/1-omdeg/medlemskap/MedlemskapConfig.tsx
@@ -1,4 +1,5 @@
-import { ISpørsmål, JaNeiSvar } from '../../../../models/spørsmal';
+import { ISpørsmål } from '../../../../models/spørsmålogsvar';
+import { JaNeiSvar } from '../../../../helpers/standardSvar';
 
 export const oppholderSegINorge: ISpørsmål = {
   søknadid: 'søkerOppholderSegINorge',
