+++ src/søknad/steg/1-omdeg/sivilstatus/begrunnelse/SøknadsBegrunnelse.tsx
@@ -9,7 +9,7 @@ import SeksjonGruppe from '../../../../../components/gruppe/SeksjonGruppe';
 import { BegrunnelseSpørsmål } from '../SivilstatusConfig';
 import { Textarea } from 'nav-frontend-skjema';
 import { FormattedHTMLMessage, useIntl } from 'react-intl';
-import { ISpørsmål } from '../../../../../models/spørsmal';
+import { ISpørsmål } from '../../../../../models/spørsmålogsvar';
 import { hentTekst } from '../../../../../utils/søknad';
 import { ISivilstatus } from '../../../../../models/steg/omDeg/sivilstatus';
 import { AlertStripeInfo } from 'nav-frontend-alertstriper';
