+++ src/søknad/steg/5-aktivitet/arbeidsforhold/OmArbeidsforholdetDitt.tsx
@@ -8,7 +8,7 @@ import SeksjonGruppe from '../../../../components/gruppe/SeksjonGruppe';
 import { Element, Undertittel } from 'nav-frontend-typografi';
 import { IAktivitet } from '../../../../models/steg/aktivitet/aktivitet';
 import { IArbeidsgiver } from '../../../../models/steg/aktivitet/arbeidsgiver';
-import { nyttTekstFelt } from '../../../../utils/søknadsfelter';
+import { nyttTekstFelt } from '../../../../helpers/tommeSøknadsfelter';
 import { hentUid } from '../../../../utils/uuid';
 
 interface Props {
@@ -17,7 +17,7 @@ interface Props {
 }
 
 const tomArbeidsgiver: IArbeidsgiver = {
-  react_key: hentUid(),
+  id: hentUid(),
   navn: nyttTekstFelt,
   arbeidsmengde: nyttTekstFelt,
 };
@@ -57,7 +57,7 @@ const OmArbeidsforholdetDitt: React.FC<Props> = ({
       </KomponentGruppe>
       {arbeidssituasjon.arbeidsforhold?.map((arbeidsgiver, index) => {
         return (
-          <SeksjonGruppe key={arbeidsgiver.react_key}>
+          <SeksjonGruppe key={arbeidsgiver.id}>
             <Arbeidsgiver
               arbeidsforhold={arbeidsforhold}
               settArbeidsforhold={settArbeidsforhold}
