+++ src/søknad/steg/5-aktivitet/underUtdanning/ErUtdanningenOffentligEllerPrivat.tsx
@@ -3,7 +3,7 @@ import MultiSvarSpørsmål from '../../../../components/spørsmål/MultiSvarSpø
 import { privatEllerOffentligSpm } from './UtdanningConfig';
 import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
 import { IUnderUtdanning } from '../../../../models/steg/aktivitet/utdanning';
-import { ISpørsmål } from '../../../../models/spørsmal';
+import { ISpørsmål } from '../../../../models/spørsmålogsvar';
 import { hentTekst } from '../../../../utils/søknad';
 import { useIntl } from 'react-intl';
 interface Props {
