+++ src/søknad/steg/5-aktivitet/underUtdanning/Utdanning.tsx
@@ -10,7 +10,7 @@ import { hentTittelMedNr } from '../../../../language/utils';
 import { Input } from 'nav-frontend-skjema';
 import { IUtdanning } from '../../../../models/steg/aktivitet/utdanning';
 import { linjeKursGrad } from './UtdanningConfig';
-import { tomPeriode } from '../../../../utils/søknadsfelter';
+import { tomPeriode } from '../../../../helpers/tommeSøknadsfelter';
 import { Undertittel } from 'nav-frontend-typografi';
 import { useIntl } from 'react-intl';
