+++ src/søknad/steg/6-meromsituasjon/SituasjonConfig.ts
@@ -1,4 +1,4 @@
-import { ISpørsmål } from '../../../models/spørsmal';
+import { ISpørsmål } from '../../../models/spørsmålogsvar';
 import {
   EDinSituasjon,
   ESagtOppEllerRedusertStilling,
